+++ rector.php
@@ -52,6 +52,8 @@ return RectorConfig::configure()
         __DIR__ . '/tests/Integration/Services/CRM/Currency',
         __DIR__ . '/src/Services/Entity/Section',
         __DIR__ . '/tests/Integration/Services/Entity/Section',
+        __DIR__ . '/src/Services/Department',
+        __DIR__ . '/tests/Integration/Services/Department',
         __DIR__ . '/tests/Unit/',
     ])
     ->withCache(cacheDirectory: __DIR__ . '.cache/rector')