+++ src/Services/CRM/CRMServiceBuilder.php
@@ -516,6 +516,23 @@ class CRMServiceBuilder extends AbstractServiceBuilder
 
         return $this->serviceCache[__METHOD__];
     }
+    
+    public function itemProductrow(): Item\Productrow\Service\Productrow
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $batch = new Item\Productrow\Batch(
+                $this->core,
+                $this->log
+            );
+            $this->serviceCache[__METHOD__] = new Item\Productrow\Service\Productrow(
+                new Item\Productrow\Service\Batch($batch, $this->log),
+                $this->core,
+                $this->log
+            );
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
 
     public function duplicate(): Duplicates\Service\Duplicate
     {